Refactor put_stream handler and authentication
Now uses a sum type to encoded a StreamingPutRequest. The first message
of the stream provides the collection ID - the rest of the messages have
no need to. Bandwidth is saved and we can no longer represent invalid
state - which is a win. Previously every message contained the
collection ID but it was ignored for all but the first message.

Authentication has also been refactored to extract an auth token from
the gRPC headers - this is implemented for all endpoints, not ony
put_stream. However, the old way of providing authentication inside the
request context is still supported so we don't break clients in the
wild.

The new authentication method is checked first - if auth token is
provided in the headers then the request context will not be examined.